+++ spec/light-service-ext/record_actions_spec.rb
@@ -5,14 +5,14 @@ module LightServiceExt
     describe '.call' do
       let(:input) { {} }
       let(:action_name) { 'some-action-name' }
-      let(:last_api_response) { nil }
+      let(:current_api_response) { nil }
       let(:invoked_action) { nil }
       let(:status) { nil }
       let(:errors) { {} }
       let(:overrides) { { errors: errors } }
       let(:ctx) { ApplicationContext.make_with_defaults(input, overrides) }
       let(:result_ctx_overrides) do
-        { status: status, errors: errors, invoked_action: invoked_action, last_api_response: last_api_response }
+        { status: status, errors: errors, invoked_action: invoked_action, current_api_response: current_api_response }
       end
       let(:result_ctx) { ApplicationContext.make_with_defaults(input, result_ctx_overrides) }
       let(:proc) { -> { result_ctx } }
@@ -29,12 +29,12 @@ module LightServiceExt
 
       context 'with api response attrs' do
         let(:invoked_action) { class_double(ApplicationAction, name: action_name) }
-        let(:last_api_response) { 'some-api-response' }
+        let(:current_api_response) { 'some-api-response' }
 
         it 'adds api_response as last api response' do
           expect(called_ctx.success?).to be_truthy
           expect(called_ctx.successful_actions).to match_array([action_name])
-          expect(called_ctx.api_responses).to match_array([last_api_response])
+          expect(called_ctx.api_responses).to match_array([current_api_response])
           expect(called_ctx.status).to eql(Status::INCOMPLETE)
           expect(called_ctx[:last_failed_context]).to be_nil
         end